+++ packages/gamut-illustrations/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.3.12](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.3.11...@codecademy/gamut-illustrations@0.3.12) (2021-01-08)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
+### [0.3.11](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.3.10...@codecademy/gamut-illustrations@0.3.11) (2021-01-08)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
+### [0.3.10](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.3.9...@codecademy/gamut-illustrations@0.3.10) (2021-01-07)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
+### [0.3.9](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.3.8...@codecademy/gamut-illustrations@0.3.9) (2021-01-07)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
 ### [0.3.8](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.3.7...@codecademy/gamut-illustrations@0.3.8) (2021-01-06)
 
 **Note:** Version bump only for package @codecademy/gamut-illustrations